+++ packages/path_provider/path_provider_macos/example/lib/main.dart
@@ -1,5 +1,5 @@
-// Copyright 2019 The Flutter Authors. All rights reserved.  Use of
-// this source code is governed by a BSD-style license that can be
+// Copyright 2019 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 // ignore_for_file: public_member_api_docs