+++ packages/image_picker/image_picker/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.6.7+21
 
-* Update the example app: remove the depreciated `RaisedButton` and `FlatButton` widgets.
+* Update the example app: remove the deprecated `RaisedButton` and `FlatButton` widgets.
 
 ## 0.6.7+20